--- android/app/src/main/java/com/nightmare/termare/VSCodePage.java
@@ -1,98 +1,0 @@
-package com.nightmare.termare;
-
-import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.graphics.Color;
-import android.net.Uri;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.Message;
-import android.util.Log;
-import android.view.Window;
-import android.view.WindowManager;
-import android.webkit.GeolocationPermissions;
-import android.webkit.WebChromeClient;
-import android.webkit.WebSettings;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-
-import androidx.annotation.NonNull;
-
-import com.nightmare.termare.R;
-
-import io.flutter.plugin.platform.PlatformPlugin;
-
-public class VSCodePage extends Activity {
-    WebView mWebView;
-    Activity context;
-
-    @SuppressLint("SetJavaScriptEnabled")
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        requestWindowFeature(Window.FEATURE_NO_TITLE);
-        setContentView(R.layout.webview);
-//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-//            Window window = getWindow();
-//            window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-//            window.setStatusBarColor(0x40000000);
-//            window.getDecorView().setSystemUiVisibility(PlatformPlugin.DEFAULT_SYSTEM_UI);
-//        }
-        //获得控件
-        context = this;
-        mWebView = (WebView) findViewById(R.id.wv_webview);
-        //访问网页
-        WebSettings mWebSettings = mWebView.getSettings();
-        //允许使用JS
-        mWebSettings.setJavaScriptEnabled(true);
-        mWebSettings.setJavaScriptCanOpenWindowsAutomatically(true);
-        mWebSettings.setUseWideViewPort(true);
-        mWebSettings.setAllowFileAccess(true);
-        // 下面这行不写不得行
-        mWebSettings.setDomStorageEnabled(true);
-        mWebSettings.setDatabaseEnabled(true);
-        mWebSettings.setAppCacheEnabled(true);
-        mWebSettings.setLoadWithOverviewMode(true);
-        mWebSettings.setDefaultTextEncodingName("utf-8");
-        mWebSettings.setLoadsImagesAutomatically(true);
-        mWebSettings.setSupportMultipleWindows(true);
-        mWebView.setWebChromeClient(webChromeClient);
-        //系统默认会通过手机浏览器打开网页，为了能够直接通过WebView显示网页，则必须设置
-        mWebView.setWebViewClient(new WebViewClient() {
-            @Override
-            public boolean shouldOverrideUrlLoading(WebView view, String url) {
-                //使用WebView加载显示url
-                view.loadUrl(url);
-                //返回true
-                return true;
-            }
-        });
-        mWebView.loadUrl("http://127.0.0.1:8080");
-    }
-
-
-    WebChromeClient webChromeClient = new WebChromeClient() {
-
-        //=========多窗口的问题==========================================================
-        @Override
-        public boolean onCreateWindow(WebView view, boolean isDialog, boolean isUserGesture, Message resultMsg) {
-            WebView childView = new WebView(context);//Parent WebView cannot host it's own popup window.
-            childView.setBackgroundColor(Color.GREEN);
-            childView.setWebViewClient(new WebViewClient() {
-                @Override
-                public boolean shouldOverrideUrlLoading(WebView view, String url) {
-                    context.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(url)));
-                    return true;
-                }
-            });
-            WebView.WebViewTransport transport = (WebView.WebViewTransport) resultMsg.obj;
-            transport.setWebView(childView);//setWebView和getWebView两个方法
-            resultMsg.sendToTarget();
-            return true;
-        }
-        //=========多窗口的问题==========================================================
-    };
-
-}
